CHANGELOG_BEGIN Moved `LedgerExportEntry` into a separate proto file. CHANGELOG_END
