+++ ledger/participant-state/kvutils/src/main/protobuf/export/ledger_export.proto
@@ -1,0 +1,39 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+syntax = "proto3";
+package com.daml.ledger.participant.state.kvutils.export;
+option java_package = "com.daml.ledger.participant.state.kvutils.export";
+option csharp_namespace = "Com.Daml.Ledger.Participant.State.KVUtils.export";
+
+import "daml_kvutils.proto";
+import "google/protobuf/timestamp.proto";
+
+// A v3 export of a KV ledger. Used by the integrity checker to test that a new version of
+// kvutils can process older ledgers.
+message LedgerExportEntry {
+  // Information about the submission.
+  message SubmissionInfo {
+    // The submitting participant ID.
+    string participant_id = 1;
+
+    // The correlation ID of the batch.
+    string correlation_id = 2;
+
+    // The submission to the ledger, wrapped in an envelope.
+    bytes submission_envelope = 3;
+
+    // the record time of the submission.
+    google.protobuf.Timestamp record_time = 4;
+  }
+
+  // A single entry, either to the log or to the state.
+  // Both key and value are serialized first.
+  message WriteEntry {
+    bytes key = 1;
+    bytes value = 2;
+  }
+
+  SubmissionInfo submission_info = 1;
+  repeated WriteEntry write_set = 2;
+}
